+++ src/Headers/CSPHeader.php
@@ -3,10 +3,10 @@
 namespace Aidantwoods\SecureHeaders\Headers;
 
 use InvalidArgumentException;
-use Aidantwoods\SecureHeaders\HeaderTemplate;
+use Aidantwoods\SecureHeaders\Header;
 use Aidantwoods\SecureHeaders\Operations\CompileCSP;
 
-class CSPHeader extends RegularHeader implements HeaderTemplate
+class CSPHeader extends RegularHeader implements Header
 {
     protected function parseAttributes()
     {
